get list product to cammand

+++ Easymakemoney/Models/ListCommand.cs
@@ -6,7 +6,7 @@ namespace Easymakemoney.Models
         public int id { get; set; }
         public string name { get; set; }
         public string photo { get; set; }
-        public float budget { get; set; }
+        public double budget { get; set; }
         public string date { get; set; }
         public int collectionId { get; set; }
     }
